% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_species.R
\name{rand_species}
\alias{rand_species}
\title{Return a list of of threatened and endangered species of n length}
\usage{
rand_species(n = 10)
}
\arguments{
\item{n}{= 10 unless specified by user}
}
\value{
A vector of threatened and endangered species of n length
}
\description{
Return a list of of threatened and endangered species of n length
}
\details{
Loads table with all threatened and endangered species data
}
