% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_species_manual.R
\name{rand_species_manual}
\alias{rand_species_manual}
\title{Return a list of of threatened and endangered species of n length}
\usage{
rand_species_manual()
}
\arguments{
\item{Prompts}{user input for an integer (n)}
}
\value{
A vector of threatened and endangered species of n length
}
\description{
Return a list of of threatened and endangered species of n length
}
\details{
Loads table with all threatened and endangered species data
}
